fix bug where voting without changing slider can fail (thanks kenny)

+++ agora/views.py
@@ -273,7 +273,7 @@ def vote_post_quick(request, topic_id, post_id):
 		print "noauth in quickvote"
 		return None
 	if request.method == 'POST':
-		voteval = int(request.POST['voteslider'])*.01
+		voteval = float(request.POST['voteslider'])*.01
 		print "QUICKVOTE", voteval
 		vote=PostVote.objects.filter(parent=post_id, author=request.user).first()
 		prnt = Post.objects.get(id=post_id)
